Make plans part of subscriptions itself
Initially the thought was to have them coming as product, but as
development happened it makes more sense to have products only have the
possible coffee beans that are going to be part of each delivery.

The products now will be tagged in a way that each subscription plan can
select the types that match the description.